requirements.txt
openai~=0.26.5
discord.py~=2.1.1
